Factored out activity tests (#344)
* Factored out activity tests

* Factored out tracer into a separate file and made it more generic in the worker rule
