+++ temporal-testing-junit4/src/main/java/io/temporal/testing/TestWorkflowRule.java
@@ -21,6 +21,7 @@ package io.temporal.testing;
 
 import io.temporal.client.WorkflowClient;
 import io.temporal.client.WorkflowClientOptions;
+import io.temporal.common.interceptors.WorkerInterceptor;
 import io.temporal.worker.Worker;
 import io.temporal.worker.WorkerFactoryOptions;
 import io.temporal.worker.WorkerOptions;
@@ -74,6 +75,7 @@ public class TestWorkflowRule implements TestRule {
   private final WorkerOptions workerOptions;
   private final boolean useExternalService;
   private final boolean doNotStart;
+  private final WorkerInterceptor[] interceptors;
 
   private String taskQueue;
 
@@ -84,7 +86,8 @@ public class TestWorkflowRule implements TestRule {
       Object[] activityImplementations,
       WorkerOptions workerOptions,
       long testTimeoutSeconds,
-      boolean doNotStart) {
+      boolean doNotStart,
+      WorkerInterceptor[] interceptors) {
     this.testEnvironment = testEnvironment;
     this.useExternalService = useExternalService;
     this.workflowTypes = workflowTypes;
@@ -92,6 +95,7 @@ public class TestWorkflowRule implements TestRule {
     this.workerOptions = workerOptions;
     this.globalTimeout = Timeout.seconds(testTimeoutSeconds);
     this.doNotStart = doNotStart;
+    this.interceptors = interceptors;
   }
 
   public static Builder newBuilder() {
@@ -110,6 +114,7 @@ public class TestWorkflowRule implements TestRule {
     private String target;
     private long testTimeoutSeconds;
     private boolean doNotStart;
+    private WorkerInterceptor[] workerInterceptors;
 
     private Builder() {}
 
@@ -129,6 +134,11 @@ public class TestWorkflowRule implements TestRule {
 
     public Builder setNamespace(String namespace) {
       this.namespace = namespace;
+      return this;
+    }
+
+    public Builder setWorkerInterceptors(WorkerInterceptor... workerInterceptors) {
+      this.workerInterceptors = workerInterceptors;
       return this;
     }
 
@@ -182,11 +192,14 @@ public class TestWorkflowRule implements TestRule {
     }
 
     public TestWorkflowRule build() {
+      namespace = namespace == null ? "UnitTest" : namespace;
+
       if (workflowClientOptions == null) {
         namespace = namespace == null ? "UnitTest" : namespace;
         workflowClientOptions = WorkflowClientOptions.newBuilder().setNamespace(namespace).build();
       }
-      WorkerFactoryOptions factoryOptions = WorkerFactoryOptions.newBuilder().build();
+      WorkerFactoryOptions factoryOptions =
+          WorkerFactoryOptions.newBuilder().setWorkerInterceptors(workerInterceptors).build();
       TestEnvironmentOptions testOptions =
           TestEnvironmentOptions.newBuilder()
               .setWorkflowClientOptions(workflowClientOptions)
@@ -203,7 +216,8 @@ public class TestWorkflowRule implements TestRule {
           activityImplementations == null ? new Object[0] : activityImplementations,
           workerOptions,
           testTimeoutSeconds == 0 ? DEFAULT_TEST_TIMEOUT_SECONDS : testTimeoutSeconds,
-          doNotStart);
+          doNotStart,
+          workerInterceptors);
     }
   }
 
@@ -224,6 +238,12 @@ public class TestWorkflowRule implements TestRule {
 
   private void shutdown() {
     testEnvironment.close();
+    if (interceptors != null) {
+      TracingWorkerInterceptor tracer = getInterceptor(TracingWorkerInterceptor.class);
+      if (tracer != null) {
+        tracer.assertExpected();
+      }
+    }
   }
 
   private void start() {
@@ -244,6 +264,18 @@ public class TestWorkflowRule implements TestRule {
   /** Returns name of the task queue that test worker is polling. */
   public String getTaskQueue() {
     return taskQueue;
+  }
+
+  /** Returns tracer. */
+  public <T extends WorkerInterceptor> T getInterceptor(Class<T> type) {
+    if (interceptors != null) {
+      for (WorkerInterceptor interceptor : interceptors) {
+        if (type.isInstance(interceptor)) {
+          return (T) interceptor;
+        }
+      }
+    }
+    return null;
   }
 
   /** Returns client to the Temporal service used to start and query workflows. */
